amharic/requirements.txt
pytesseract
Pillow
matplotlib
opencv-python